src/pyproject.toml
[tool.poetry]
name = "cpython-energy-impact-experiment"
version = "0.1.0"
description = ""
authors = ["Doubleblinder <double@blind.er>"]
readme = "README.md"
packages = [{include = "cpython_energy_impact_experiment"}]

[tool.poetry.dependencies]
python = "^3.10"
otii_tcp_client_python = { path = "./dependencies/otii_tcp_client_python-2.2.5-py3-none-any.whl" }
appdirs = "^1.4.4"
requests = "^2.31.0"
httpx = "^0.26.0"
pandas = "^2.2.0"


[tool.poetry.group.dev.dependencies]
ipython = "^8.21.0"
notebook = "^7.1.1"
scipy = "^1.12.0"
matplotlib = "^3.8.3"
pyarrow = "^15.0.1"
rise = "^5.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
